Separate class TextEditor into independent file
