+++ gui/src/class_TextEditor.h
